Implement filtering of nulls, defaults and optional clauses
All nullish values (null, undefined, empty string, but not false)
are now replaced with a user-specified string, or '' by default.
Furthermore, each placeholder can optionally specify the clause
to use when the value is nullish, and that clause can have placeholders
too.
If all replacements in a phrase are nullish, the replacement fails.

The phrase can now also contain optional clauses. They are replaced
as normal, but if any value in the optional clause is nullish, the
whole phrase silently expands to the empty string.